Implement configure services screen for Styist App
The screen allows choosing the template set first,
then adding/removing/editing the services in the template.

The user can also start with a blank list of services.

Also removed StoreService and related code which was
unneccessary. We are going to use a different state
management approach later (likely ngrx).