Create simplistic heuristic dynamic pricing calculation function
This is in preparation to start calculating the price for a client
for a given service and given date ranges.

For full description of functionality see documentation of
calc_client_prices() function.

Also added generation of html coverage report in Makefile. I am going
to try and see if VSTS can display it once this commit is merged.