Replace non-existing functionality with existing in footer
The History and Invite screens are not ready. I temporarily replaced
these 2 buttons by buttons that navigate to working hours and profile
for now to avoid having broken functionality in the app. We cannot
have broken functionality in the app because it will not pass Apple's
review.

Also added a flag to avoid printing environment-related webpack
messages more than once during build.