Start implementation of Client Invitations screen
We get the full list of contacts from the phone and allow the user to
select the phone numbers that they want to send invitations to.
The list is sorted alphabetically and grouped by first letter.

There is a search capability that allows searching by name or phone
number using fuse.js fuzzy searching library. Found matches are
highlighted. When searching the grouping by first letter is disabled
and the matches are sorted by score.

User can add a phone number manually to the list too.

Note: this is an incomplete feature. The design is not ready and
functionality may also need to be extended when we have the final
design. However, this feature is hidden using a feature flag so
it is acceptable to merge this change as is.

I also fixed the following (not directly related, but annoying):

1. Added server status bar to Login/Register form. We need this on all
screens that do no handle server availability explicitly.

2. Change BB_ENV variable to MB_ENV. BB_ENV will be honored for now
for backwards compatibilty. Will remove once all parties are informed.

Github issue: #449