Add invitation states to the Invitations screen
Load invitation states from the server and show the states
in the list of contacts.

Validate phone numbers and only allow to send invitations to
valid phone numbers. A number is considered valid if it is
a valid phone number in the US or is an international phone
number starting with + (plus) sign.

Fix reading contacts on iOS and add ability to re-request
contacting reading permission.

Skin over the Invitations screen (done by Anahit) according
to the new design.