Remove Facebook login from Stylist App (#630)
The goal is to have phone-based login in the near future. Facebook
login currently has bugs. We decided to remove it completely. Fixing
Facebook login now is a waste of time since we will not need it anymore
once phone login is implemented. (I think phone login is much nicer and
Facebook login simply is unnecessary once it is implemented).

Discussed here:
https://madebeauty.slack.com/archives/CA1GS4X6J/p1534540999000100