Add rebooking capability
I also added ability to hide alerts of any errors from API responses.
I needed this in this commit first but then implemented without it,
but I am keeping this ability since it is generic and can be useful
in the future.

Also moved base-api-service.ts to core directory of stylist since it
is not a shared code.