[HOTFIX] Replace deep-equal with fast-deep-equal to fix the build
The build was failing on Windows because of 'require' statement.
Instead of tryring to figure out what's wrong with it I deciced to
get rid of it and use the nice fast-deep-equal package which plays
niceley with TypeScript import.

It fixes the build and fast-deep-equal is also about 4x
faster.