Reducity severity of ServerUnreachableError to warning
The server can be unreachable for a number of reasons that are
not our fault (e.g. network problems on the user side). We don't
want to be alerted every time this happens.

We still want this to be reported to Sentry as a warning so that
we can set rules to trigger alerts when we have sudden spikes of
ServerUnreachableError which can indicate our server is down.

JIRA: FRON-907