Add nice alert when "Loading chunk failed" happens
We do not yet know what is the root cause for "Loading chunk failed".
While we are investigating the issue I am adding an nicer alert
message with a suggstion to restart the app when this error happens.
This is a temporay solution until we fix the root cause.

JIRA: FRON-919