Disable scrolling to top on slide changes in Stylist App
The scrolling to top is triggering an unknown internal Ionic
bug. At the same time it highly arguable that scrolling
to top is the deirable behavior. A better behavior would be
if each slided remembered its scroll position and we didn't
try to scroll to top on slide changes.

I am disabling this scrolling for now to get rid of the bug.
Later we can think about changing the slide change / scroll
functionality the way that we want. For now this is a good
enough fix.

JIRA: https://madebeauty.atlassian.net/browse/FRON-962