Only show "cancel" option for new appointments in Client App
Previously "cancel" was also shown for no-show appointments
which would result in error message because it is not allowed
by backend (and makes no sense).

Note: in the future we need to handle no-show better in the
Client App, e.g. indicate it somehow, but that's a separate
task.

JIRA: FRON-1053