[HOTFIX] Explicitly add release id to Sentry error
In Sentry event payload created by Cordova, release and dist ids can
be set in two places - on upper level (eg. `event.releaseName`) and
in the `extra` dictionary in the form of `__sentry_release` and
`__sentry_dist` fields. iOS library expects the latter from the event
payload received from Cordova, so it's necessary to set those fields
explicitly. To do that through sentry-cordova, we need to call
.setRelease() together with .setDist().