Cleanup auth in Stylist App
We had a messy situation with Auth in the apps which is a result
of diverged auth implementations and later not fully done merging.

This change finalizes the merging and makes Client and Stylist App
authentication code uniform and almost entirely shared.

The only part that is now different is email-based auth that is
still used by Client App E2E tests, which will be converted to
phone auth in a separate PR later.

This cleanup was done when I found that we do not send userId to
Google Analytics. It turned out that this functionality was broken
since last changes to auth. This change prepares a fixe for that bug too:
JIRA: https://madebeauty.atlassian.net/browse/FRON-1074

However, for userId to be reported to GA we also need backend to
start sending user_uuid in the response to certain API. It is
a separate backend task in JIRA: https://madebeauty.atlassian.net/browse/BAC-314