Extract AppStorage as independent class (#1030)
This was previously Stylist App specific class.

We will need to use a similar persistent storage in the Client App
in the near future. I extract a generic AppStorage class and
now we have specialized StylistAppStorage class that extends it
and defines the actual data that needs to be stored. We will be able
to easy do the same for Client App later.