Persistently remember that push permission was denied
If push permission was denied we previously we were not remembering
the fact and after 14 days would try to obtain permission again.
We would show the permission priming screen however clicking
"Allow Notification" button would silently fail because denials
are permanent under iOS.

Now we remember the denials in persistent storage and do not attempt
to show the permission priming screen again anymore.