Ensure "Welcome to MADE!" screen is not shown for completed profiles on App start
We had a bug on app startup. We were not refreshing the locally stored auth
data which includes profile status. The profile status is used to know which
screen must be shown at the start of the app.

Now we always refresh the auth on startup of Client App (like we were already
doing for Stylist App).

JIRA: https://madebeauty.atlassian.net/browse/FRON-1133