Add defenses again locally saved auth data being incorrect in Client App
If the auth data is incorrect (e.g. profileStatus is missing,
which is possible because in the past we didn't save it) then
we need to still work as best as we can.

We now either completely discard locally saved auth and
as a result of it require re-login or will assume the profile
is incomplete and will show the corresponding screen (the exact
action depends on where exactly and how we detect incorrectness
in the saved data).

JIRA: https://madebeauty.atlassian.net/browse/FRON-1147