Introduce Made Analytics
We now send simple analytics data to our own backend.
This currently mirrors the trackView event that we send to
Google Analytics.

The implementation is very simple: there are no retries,
no queuing of requests if the backend is not available.

This will make it easy to do certain types of analysis using
our own data.