Update cordova-plugin-ionic-webview to latest version (#1165)
This is needed to solve the "white screen on startup" bug.
This is a hard to reproduce bug but it seems extremelly similar
to this bug description in cordova-plugin-ionic-webview:
https://github.com/ionic-team/cordova-plugin-ionic-webview/pull/45

The latest version of cordova-plugin-ionic-webview contains
a fix for this bug.

Unfortunately I have no way of reproducing the bug and seeing
if this actually fixes it. However, worst case we will just
have a fresher version of cordova-plugin-ionic-webview which
supposedly also brings performance improvements, so even if
that particular bug is not fixed, this is still a positive change.

We set <preference name="UseScheme" value="true" /> so it won't
use the webserver port on iOS 11+. It will still use it on iOS 10
as the web server is still used on iOS 10 apps. This port number
number is specified using WKPort parameter and is recommended
to use an unique port for iOS 10 users. For more details see:
https://github.com/ionic-team/cordova-plugin-ionic-webview/issues/165#issuecomment-445200469

Also, disable auto-hiding of splash screen and make sure it is
hidden at the very end of applicaiton startup. Previously the
auto-hide could hide the splash screen after 3 seconds which
sometimes was not enough and would lead to a blank screen until
the first screen was shown.

JIRA: https://madebeauty.atlassian.net/browse/FRON-1180