Add deeplinking to Client and Stylist Apps
Links in the form of https://madebauty.com/get/ will open
Client App.

Links in the form of https://madebauty.com/getpro/ will open
Stylist App.

Note: it seems like the concept of "url prefix" does not work
properly in Ionic deeplink plugin on Android (I didn't verify iOS yet)
and it only uses the domain name to identify which app to open so
if you have both apps installed then it will ask which one
to open and does not differentiate properly based on the url (/get vs /getpro).

I do not want too much waste time and investigste why the plugin
is not doing what it promises to do since this is an edge case
that we don't care much about.

For now it is enough that it reacts correctly to links in the form of
https://madebeauty.com/

I will need to verify iOS build after it is ready in TestFlight.